requirements.txt
numpy
matplotlib
librosa
pandas
torch
torchvision
sklearn
ipywidgets
pytorch-ignite
tqdm
visdom
future
tb-nightly